refactor: extract Module components in ffc-volunteer proving ground
Extracts the 7 hardcoded Modulecard blocks from
src/components/ffc-volunteer-proving-ground-core-competencies/Modules-section/index.tsx
into individual component files (Module1.tsx through Module7.tsx).
This significantly reduces the complexity and length of the index file
(from ~500 lines to ~22 lines), making the codebase easier to read and maintain.
